search permissions: reduce findDvObject calls #2036
The profiler shows significantly better performance when we reduce the
number of calls to findDvObject so we pass around the DvObject itself
more often, rather than looking it up over and over.

The profiler also suggests that batching the commits to Solr may also
yield a good performance gain but that's for a different commit.

The index API endpoint "missing" is removed as part of this commit since
it relies on code that was rewritten as part of this commit but it was
little used anyway since there's a newer "skipIndexed" boolean for other
index endpoints for catching up an index. We can revisit if we want to
re-add the "missing" endpoint.

Finally the createSolrDoc method was used to SearchUtil and tests were
written for it.

Here are the numbers for before and after this commit from running
scripts/issues/2036/grant-role-then-revoke

        | Solr Up   | Solr Down
------- | --------- | -------
grant   | 0m25.653s | 0m9.699s
revoke  | 0m23.539s | 0m9.968s

        | Solr Up   | Solr Down
------- | --------- | -------
grant   | 0m7.422s  | 0m10.348s
revoke  | 0m6.758s  | 0m10.341s